#30882 enhance CSV import functionality with improved error handling and logging
